+++ benchmarks/fogbench/templates/fogbench_sensor_coap.template.json
