+++ src/python/.gitignore
@@ -53,6 +53,3 @@ pylint_*.log
 .coverage
 htmlcov/
 coverage.xml
-
-# temporary
-out/
